Zod schemas for payment gateways
