+++ modules/zuora/src/account.ts
@@ -1,7 +1,5 @@
-import type {
-	PaymentGateway,
-	PaymentMethod,
-} from '@modules/zuora/orders/paymentMethods';
+import type { PaymentGateway } from '@modules/zuora/orders/paymentGateways';
+import type { PaymentMethod } from '@modules/zuora/orders/paymentMethods';
 import type { ZuoraAccount } from './types';
 import { voidSchema, zuoraAccountSchema } from './types';
 import type { ZuoraClient } from './zuoraClient';
